MOD:Merged the vehicles routes and middleware
