--- api/src/middlewares/vehiclePost.js
@@ -1,8 +1,0 @@
-const axios = require("axios");
-
-const vehiclePost = async (data) => {
-    const details = {post : "se posteo:", data}
-    return details;
-}
-
-module.exports = vehiclePost;
